+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * Add dummy hooks for pulseaudio
 * Add versions for SteamUserStats API
 * Add the setcap command in hook so that it is called on make install
+* Add support to .AppImage games
 
 ### Changed